Bugt with lombok annotations

+++ procrapi/src/main/java/org/bem/procrapi/ProcrapiApplication.java
@@ -23,7 +23,7 @@ public class ProcrapiApplication implements CommandLineRunner {
         String gestionnaire_mail = "dali.mabrouk@miage.fr";
         if (repositoryUtilisateur.findByEmail(gestionnaire_mail).isEmpty()) {
             Utilisateur gestionnaireDuTempsPerdu = new Utilisateur();
-            gestionnaireDuTempsPerdu.setId(1);
+            gestionnaireDuTempsPerdu.setId((long)1);
             gestionnaireDuTempsPerdu.setRole(RoleUtilisateur.GESTIONNAIRE_DU_TEMPS_PERDU);
             gestionnaireDuTempsPerdu.setEmail(gestionnaire_mail);
             repositoryUtilisateur.save(gestionnaireDuTempsPerdu);
